+++ visualization/app/app.module.ts
@@ -13,11 +13,37 @@ import { AttributeSideBarModule } from "./codeCharta/ui/attributeSideBar/attribu
 import { AttributeSideBarComponent } from "./codeCharta/ui/attributeSideBar/attributeSideBar.component"
 import { Export3DMapButtonComponent } from "./codeCharta/ui/export3DMapButton/export3DMapButton.component"
 import { Export3DMapButtonModule } from "./codeCharta/ui/export3DMapButton/export3DMapButton.module"
+import { LabelledColorPickerComponent } from "./codeCharta/ui/labelledColorPicker/labelledColorPicker.component"
+import { LegendPanelComponent } from "./codeCharta/ui/legendPanel/legendPanel.component"
+import { LegendPanelModule } from "./codeCharta/ui/legendPanel/legendPanel.module"
+import { ColorPickerForMapColorModule } from "./codeCharta/ui/colorPickerForMapColor/colorPickerForMapColor.module"
+import { ColorPickerForMapColorComponent } from "./codeCharta/ui/colorPickerForMapColor/colorPickerForMapColor.component"
+import { MarkFolderColorPickerModule } from "./codeCharta/ui/nodeContextMenu/markFolderColorPicker/markFolderColorPicker.module"
+import { MarkFolderColorPickerComponent } from "./codeCharta/ui/nodeContextMenu/markFolderColorPicker/markFolderColorPicker.component"
 
 @NgModule({
-	imports: [BrowserModule, UpgradeModule, MapTreeViewModule, MatchingFilesCounterModule, AttributeSideBarModule, Export3DMapButtonModule],
+	imports: [
+		BrowserModule,
+		UpgradeModule,
+		MapTreeViewModule,
+		MatchingFilesCounterModule,
+		AttributeSideBarModule,
+		Export3DMapButtonModule,
+		LegendPanelModule,
+		ColorPickerForMapColorModule,
+		MarkFolderColorPickerModule
+	],
 	declarations: [],
-	entryComponents: [MapTreeViewComponent, MatchingFilesCounterComponent, AttributeSideBarComponent, Export3DMapButtonComponent]
+	entryComponents: [
+		MapTreeViewComponent,
+		MatchingFilesCounterComponent,
+		AttributeSideBarComponent,
+		Export3DMapButtonComponent,
+		LegendPanelComponent,
+		LabelledColorPickerComponent,
+		ColorPickerForMapColorComponent,
+		MarkFolderColorPickerComponent
+	]
 })
 export class AppModule {
 	constructor(@Inject(UpgradeModule) private upgrade: UpgradeModule) {}