+++ visualization/app/codeCharta/state/store/fileSettings/markedPackages/markedPackages.reducer.spec.ts
@@ -1,5 +1,5 @@
 import { markedPackages } from "./markedPackages.reducer"
-import { MarkedPackagesAction, setMarkedPackages, unmarkPackage } from "./markedPackages.actions"
+import { MarkedPackagesAction, setMarkedPackages, calculateMarkedPackages, unmarkPackage } from "./markedPackages.actions"
 import { MARKED_PACKAGES } from "../../../../util/dataMocks"
 
 describe("markedPackages", () => {
@@ -30,6 +30,13 @@ describe("markedPackages", () => {
 			const result = markedPackages([MARKED_PACKAGES[0]], unmarkPackage(0))
 
 			expect(result).toEqual([])
+		})
+	})
+
+	describe("Action: CALCULATE_MARKED_PACKAGES", () => {
+		it("should add all packages", () => {
+			const result = markedPackages([MARKED_PACKAGES[0]], calculateMarkedPackages(MARKED_PACKAGES.slice(1)))
+			expect(result).toEqual(MARKED_PACKAGES)
 		})
 	})
 })